About us added for the screen
