+++ lib/navigationBarScreens/profile_screen.dart
@@ -1,9 +1,12 @@
 
+import 'package:else_app_two/profileTab/profile_screen_route.dart';
 import 'package:flutter/material.dart';
 
 class Profile extends StatelessWidget {
 
   final List<String> entries = <String>['Help', 'About', 'Share', 'Rate', 'Logout'];
+
+  ProfileScreenRoute handler = new ProfileScreenRoute();
 
   @override
   Widget build(BuildContext context) {
@@ -14,12 +17,7 @@ class Profile extends StatelessWidget {
         return ListTile(
           title: Center(child: Text('${entries[index]}')),
           onTap: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(
-                builder: (context) => AboutUsDetailScreen(index: index),
-              ),
-            );
+              handler.routeToProfileOptions(context, index);
           },
         );
       },
@@ -27,24 +25,3 @@ class Profile extends StatelessWidget {
   }
 }
 
-class AboutUsDetailScreen extends StatelessWidget {
-  // Declare a field that holds the Todo.
-  final int index;
-
-  // In the constructor, require a Todo.
-  AboutUsDetailScreen({Key key, @required this.index}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    // Use the Todo to create the UI.
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('About US'),
-      ),
-      body: Padding(
-        padding: EdgeInsets.all(16.0),
-        child: Text('Kuch bhi likh diya hai bus'),
-      ),
-    );
-  }
-}
